refactor: consolidate entry-point discovery on the fallow-core implementation
Delete the engine's forked entry-point discovery and route the engine's
discover wrappers through core_backend pass-throughs to the fallow-core
implementation, mirroring the discovery-walk consolidation.
BackendAggregatedPluginResult now stores the core AggregatedPluginResult
directly, so plugin entry-point discovery needs no mirror-type
reconstruction and keeps per-field pattern provenance. The engine's three
end-to-end entry-point tests move to fallow-core, the engine drops its
now-unused regex dependency, and the boundary allowlist gains the three
core discovery calls. The skipped-entry warning dedupe becomes a single
process-wide set instead of one per crate.
